Expand on the C compiler/ruby headers requirement for mysql::ruby
That mysql::ruby executes at compile-time tripped me up for several hours yesterday. Until users become familiar with the anatomy of a Chef run (probably by encountering a situation like this), it's not immediately clear that it's even possible for recipes to (essentially) run in a different order than the one they're specified in. This change expands on the existing warning in the readme to make the mysql::ruby preconditions clearer, and provides an example of how to do that with build-essential.